add: support postgresql dialect
