+++ _example/pg/postgresql.sql
@@ -1,0 +1,52 @@
+-- generated by github.com/mackee/go-genddl. DO NOT EDIT!!!
+
+CREATE TABLE "location" (
+    "id" BIGINT NOT NULL PRIMARY KEY GENERATED BY DEFAULT AS IDENTITY,
+    "name" VARCHAR NOT NULL,
+    "description" VARCHAR NOT NULL,
+    "place" BYTEA NOT NULL
+) ;
+
+
+CREATE TABLE "product" (
+    "id" BIGINT NOT NULL PRIMARY KEY GENERATED BY DEFAULT AS IDENTITY,
+    "name" VARCHAR NOT NULL,
+    "type" BIGINT NOT NULL,
+    "user_id" BIGINT NOT NULL,
+    "received_user_id" BIGINT NULL,
+    "description" TEXT NOT NULL,
+    "full_description" VARCHAR NOT NULL,
+    "size" INTEGER NULL,
+    "status" SMALLINT NOT NULL,
+    "category" SMALLINT NOT NULL,
+    "created_at" TIMESTAMP(6) WITH TIME ZONE NOT NULL,
+    "updated_at" TIMESTAMP(6) WITH TIME ZONE NULL,
+    "removed_at" TIMESTAMP(6) WITH TIME ZONE NULL,
+    UNIQUE ("user_id", "type")
+) ;
+CREATE INDEX product_user_id_created_at ON product ("user_id", "created_at");
+
+
+CREATE TABLE "user" (
+    "id" BIGINT NOT NULL PRIMARY KEY GENERATED BY DEFAULT AS IDENTITY,
+    "name" VARCHAR(255) NOT NULL UNIQUE,
+    "age" BIGINT NULL,
+    "message" VARCHAR NULL,
+    "icon_image" BYTEA NOT NULL,
+    "created_at" TIMESTAMP WITH TIME ZONE NOT NULL,
+    "updated_at" TIMESTAMP WITH TIME ZONE NULL
+) ;
+
+CREATE VIEW "user_product"
+  ("u_id", "u_name", "ru_name", "p_id", "p_type") AS
+  SELECT u.id, u.name, ru.name, p.id, p.type FROM product AS p
+    INNER JOIN "user" AS u ON p.user_id = u.id
+    LEFT JOIN "user" AS ru ON p.received_user_id = ru.id;
+
+CREATE VIEW "user_product_structured"
+  ("p_id", "p_name", "p_type", "p_user_id", "p_received_user_id", "p_description", "p_full_description", "p_size", "p_status", "p_category", "p_created_at", "p_updated_at", "p_removed_at", "u_id", "u_name", "u_age", "u_message", "u_icon_image", "u_created_at", "u_updated_at", "ru_id", "ru_name", "ru_age", "ru_message", "ru_icon_image", "ru_created_at", "ru_updated_at") AS
+  SELECT p.*, u.*, ru.* FROM product AS p
+    INNER JOIN "user" AS u ON p.user_id = u.id
+    LEFT JOIN "user" AS ru ON p.received_user_id = ru.id;
+
+ALTER TABLE "product" ADD CONSTRAINT "fk_product_user_id" FOREIGN KEY ("user_id") REFERENCES "user"("id") ON DELETE CASCADE ON UPDATE CASCADE;
