+++ postgresql.go
@@ -6,66 +6,72 @@ import (
 	"github.com/mackee/go-genddl/index"
 )
 
-const (
-	MYSQL_DEFAULT_VARCHAR_SIZE = "191"
-)
-
-type MysqlDialect struct {
+type PostgresqlDialect struct {
 	Collate string
 }
 
-func (m MysqlDialect) DriverName() string { return "mysql" }
+func (m PostgresqlDialect) DriverName() string { return "pg" }
 
-func (m MysqlDialect) ToSqlType(col *ColumnMap) (string, error) {
+func (m PostgresqlDialect) ToSqlType(col *ColumnMap) (string, error) {
 	column := ""
+	if v, ok := col.TagMap["def"]; ok {
+		return v, nil
+	}
 
 	switch col.TypeName {
 	case "bool":
 		column = "BOOLEAN"
-	case "int8":
-		column = "TINYINT"
-	case "uint8":
-		column = "TINYINT unsigned"
-	case "int", "int16", "int32":
+	case "int8", "uint8":
+		column = "SMALLINT"
+	case "int", "int16", "int32", "uint", "uint16":
 		column = "INTEGER"
-	case "uint16", "uint32":
-		column = "INTEGER unsigned"
-	case "int64":
+	case "uint32", "int64":
 		column = "BIGINT"
 	case "uint64":
-		column = "BIGINT unsigned"
+		column = "NUMERIC(20)"
 	case "float64":
-		column = "DOUBLE"
+		column = "DOUBLE precision"
 	case "float32":
-		column = "FLOAT"
+		column = "REAL"
 	case "string":
 		if _, ok := col.TagMap["text"]; ok {
 			column = "TEXT"
-		} else if _, ok := col.TagMap["mediumtext"]; ok {
-			column = "MEDIUMTEXT"
 		} else {
-			size := MYSQL_DEFAULT_VARCHAR_SIZE
-			if v, ok := col.TagMap["size"]; ok {
-				size = v
+			tname := "VARCHAR"
+			if v, ok := col.TagMap["type"]; ok {
+				tname = v
 			}
-			column = "VARCHAR(" + size + ")"
+			if size, ok := col.TagMap["size"]; ok {
+				column = tname + "(" + size + ")"
+			} else {
+				column = tname
+			}
 		}
 		if m.Collate != "" {
 			column += " COLLATE " + m.Collate
 		}
 	case "time.Time":
-		column = "DATETIME"
+		column = "TIMESTAMP"
 		if v, ok := col.TagMap["precision"]; ok {
 			column += "(" + v + ")"
+		}
+		if _, ok := col.TagMap["withouttimezone"]; ok {
+			column += " WITHOUT TIME ZONE"
+		} else {
+			column += " WITH TIME ZONE"
 		}
 	case "[]byte":
 		if v, ok := col.TagMap["type"]; ok {
 			column = v
 		} else {
-			column = "BLOB"
+			column = "BYTEA"
 		}
 	default:
-		return "", fmt.Errorf("unsupported type: %s, column=%s", col.TypeName, col.Name)
+		if v, ok := col.TagMap["type"]; ok {
+			column = v
+		} else {
+			return "", fmt.Errorf("unsupported type: %s, column=%s", col.TypeName, col.Name)
+		}
 	}
 
 	if _, ok := col.TagMap["null"]; ok || col.IsNullable {
@@ -74,9 +80,6 @@ func (m MysqlDialect) ToSqlType(col *ColumnMap) (string, error) {
 		column += " NOT NULL"
 	}
 
-	if v, ok := col.TagMap["srid"]; ok {
-		column += " SRID " + v
-	}
 	if v, ok := col.TagMap["default"]; ok {
 		column += " DEFAULT " + v
 	}
@@ -87,24 +90,21 @@ func (m MysqlDialect) ToSqlType(col *ColumnMap) (string, error) {
 		column += " PRIMARY KEY"
 	}
 	if _, ok := col.TagMap["autoincrement"]; ok {
-		column += " AUTO_INCREMENT"
+		column += " GENERATED BY DEFAULT AS IDENTITY"
 	}
 
 	return column, nil
 }
 
-func (m MysqlDialect) CreateTableSuffix() string {
-	if m.Collate != "" {
-		return fmt.Sprintf("ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=%s", m.Collate)
-	}
-	return "ENGINE=InnoDB DEFAULT CHARSET=utf8mb4"
+func (m PostgresqlDialect) CreateTableSuffix() string {
+	return ""
 }
 
-func (m MysqlDialect) QuoteField(field string) string {
-	return "`" + field + "`"
+func (m PostgresqlDialect) QuoteField(field string) string {
+	return `"` + field + `"`
 }
 
-func (m MysqlDialect) ForeignKey(option index.ForeignKeyOption) string {
+func (m PostgresqlDialect) ForeignKey(option index.ForeignKeyOption) string {
 	switch option {
 	case index.ForeignKeyDeleteRestrict:
 		return "ON DELETE RISTRICT"
